requirements.txt
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==2.1.1
click==8.1.8
colorama==0.4.6
fastapi==0.95.1
flake8==7.2.0
Flask==2.3.3
Flask-Cors==4.0.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
mysql-connector-python==9.3.0
packaging==25.0
pycodestyle==2.13.0
pydantic==1.10.7
pyflakes==3.3.2
python-dotenv==0.21.0
requests==2.31.0
sniffio==1.3.1
starlette==0.26.1
typing_extensions==4.13.1
urllib3==1.26.20
uvicorn==0.23.1
Werkzeug==3.1.3
Flask
mysql-connector-python
python-dotenv
gunicorn
